Add --cpu-temp-fix flag to fix high CPU usage
Add a --cpu-temp-fix flag that is used as a temporary fix to deal with
a tight loop in the main_loop function until a better solution can be
found.  The --cpu-temp-fix flag adds a 10ms wait time at the end of the
while loop to slow the loop down enough to use a small fraction of the
CPU usage it was using before.  As the --cpu-temp-fix flag is only a
temporary workaround to the high CPU usage issue, it is not known at
this moment if having the flag set by default will ruin any payloads
that rely on very high data rates, so it is not set by default.  It is
recommended for most runs to set the --cpu-temp-fix flag.